+++ cbor/src/main/java/com/fasterxml/jackson/dataformat/cbor/CBORParser.java
@@ -959,7 +959,7 @@ public class CBORParser extends ParserMinimalBase
             case 25: // 16-bit float...
                 // As per [http://stackoverflow.com/questions/5678432/decompressing-half-precision-floats-in-javascript]
                 {
-                    _numberFloat = (float) _decodeHalfSizeFloat();
+                    _numberFloat = _decodeHalfSizeFloat();
                     _numTypesValid = NR_FLOAT;
                 }
                 return (_currToken = JsonToken.VALUE_NUMBER_FLOAT);
@@ -1016,7 +1016,7 @@ public class CBORParser extends ParserMinimalBase
                         long unsignedBase = (long) i & 0xFFFFFFFFL;
                         l = -unsignedBase - 1L;
                     } else {
-                        l = (long) i;
+                        l = i;
                         l = l & 0xFFFFFFFFL;
                     }
                     return String.valueOf(l);
@@ -2097,7 +2097,7 @@ public class CBORParser extends ParserMinimalBase
     protected void convertNumberToLong() throws IOException
     {
         if ((_numTypesValid & NR_INT) != 0) {
-            _numberLong = (long) _numberInt;
+            _numberLong = _numberInt;
         } else if ((_numTypesValid & NR_BIGINT) != 0) {
             if (BI_MIN_LONG.compareTo(_numberBigInt) > 0
                     || BI_MAX_LONG.compareTo(_numberBigInt) < 0) {
@@ -2173,13 +2173,13 @@ public class CBORParser extends ParserMinimalBase
         if ((_numTypesValid & NR_BIGDECIMAL) != 0) {
             _numberDouble = _numberBigDecimal.doubleValue();
         } else if ((_numTypesValid & NR_FLOAT) != 0) {
-            _numberDouble = (double) _numberFloat;
+            _numberDouble = _numberFloat;
         } else if ((_numTypesValid & NR_BIGINT) != 0) {
             _numberDouble = _numberBigInt.doubleValue();
         } else if ((_numTypesValid & NR_LONG) != 0) {
             _numberDouble = (double) _numberLong;
         } else if ((_numTypesValid & NR_INT) != 0) {
-            _numberDouble = (double) _numberInt;
+            _numberDouble = _numberInt;
         } else {
             _throwInternal();
         }
